requirements.txt
Django==3.0.5
Werkzeug==1.0.1
django-environ==0.4.5
serverless-psycopg2==0.0.2
djangorestframework==3.11.0
drf-yasg==1.17.1
requests==2.22.0
django-filter==2.2.0
django-cors-headers==3.2.1
django-extensions==2.2.9
bpython==0.19
